+++ org.gitools.matrix/src/main/java/org/gitools/matrix/transform/parameters/AggregatorParameter.java
@@ -1,0 +1,12 @@
+package org.gitools.matrix.transform.parameters;
+
+
+import org.gitools.api.analysis.IAggregator;
+
+public class AggregatorParameter extends AbstractFunctionParameter<IAggregator> {
+
+    @Override
+    public boolean validate(IAggregator parameter) {
+        return super.validate(parameter);
+    }
+}