+++ org.gitools.matrix/src/main/java/org/gitools/matrix/transform/parameters/DoubleParameter.java
@@ -1,0 +1,18 @@
+package org.gitools.matrix.transform.parameters;
+
+
+public class DoubleParameter extends AbstractFunctionParameter<Double> {
+
+    @Override
+    public boolean validate(Double parameter) {
+
+        super.validate(parameter);
+
+        if (parameter.isNaN() && parameter.isInfinite()) {
+            return false;
+        }
+
+        return true;
+    }
+
+}